day 3 2025

+++ AoC-2025/src/main/java/me/imspooks/adventofcode/v2025/DayTemplate.java
@@ -17,12 +17,12 @@ public class DayTemplate implements Day {
     }
 
     @Part(part = 1)
-    public int firstPart() {
+    public long firstPart() {
         return 0;
     }
 
     @Part(part = 2)
-    public int secondPart() {
+    public long secondPart() {
         return 0;
     }
 }
